+++ examples/webpack/assets/scripts/ts/greet.ts
@@ -1,9 +1,7 @@
-console.log('This is sample1');
-
-export class Greeter {
-    constructor(public greeting: string) { }
+export default class Greeter {
+    constructor() { }
     greet() {
       let square = (x:number)=>x*x;
-      return `Hello, TypeScript! num=${square(2)}}`;
+      return `Hello, TypeScript! num=${square(2)}`;
     }
 }