+++ src/plugins/TypeScriptPlugin.ts
@@ -3,7 +3,7 @@
  */
 import * as gulp from 'gulp';
 import * as upath from 'upath';
-import {is, msg, toPromise, warn} from '../utils/utils';
+import {dmsg, is, msg, toPromise, warn} from '../utils/utils';
 import {GulpStream, Options} from "../core/types";
 import {GBuilder} from "../core/builder";
 import {GPlugin} from "../core/plugin";
@@ -15,17 +15,27 @@ export class TypeScriptPlugin extends GPlugin {
     const tsOpts = builder.moduleOptions.typescript || {};
     const tsConfig = builder.buildOptions.tsConfig;
 
+    // normalize outDir and outFile
+    let outFile = (tsOpts.output && tsOpts.output.filename) || builder.conf.outFile;
+    let outDir = (tsOpts.output && tsOpts.output.path) || builder.conf.dest;
+    if (outDir && outFile && !upath.isAbsolute(outFile)) outFile = upath.join(outDir, outFile);
+    if (outFile) {
+      outDir = upath.dirname(outFile);
+      outFile = upath.basename(outFile)
+    }
+
     // conf setting will override module settings(mopts)
-    if (builder.conf.outFile) Object.assign(tsOpts, {outFile:upath.resolve(builder.conf.outFile)});
-    if (builder.conf.outFile && is.String(builder.conf.dest))
-      tsOpts.outDir = upath.resolve(builder.conf.dest as string);
+    if (outFile) tsOpts.outFile = outFile;
+    if (outDir) tsOpts.outDir = outDir;
 
     // check lint option
     if (builder.buildOptions.lint) {
       const tslint = require('gulp-tslint');
-      const tslintOpts = builder.moduleOptions.tslint || {};
-      msg('tslintOpts =', tslintOpts);
-      builder.pipe(tslint(tslintOpts)).pipe(tslint.report(tslintOpts.report));
+      // const tslintOpts = builder.moduleOptions.tslint || {formatter: 'stylish'};
+      const tslintOpts = builder.moduleOptions.tslint || {formatter: 'stylish'};
+      const reportOpts = tslintOpts.report || {};
+      // dmsg('[TypeScriptPlugin]tslint Options =', tslintOpts, reportOpts);
+      builder.pipe(tslint(tslintOpts)).pipe(tslint.report(reportOpts));
     }
 
     const typescript = require('gulp-typescript');
@@ -33,21 +43,16 @@ export class TypeScriptPlugin extends GPlugin {
     if (tsConfig) {
       try {
         tsProject = typescript.createProject(tsConfig, tsOpts);
-        if (builder.buildOptions.printConfig) {
-          msg(`[TypeScriptPlugin]tsconfig evaluated(buildName:${builder.conf.buildName}):\n`,
-            Object.assign({}, tsProject.config, {"compilerOptions":tsOpts}));
-        }
       }
       catch (e) {
         if (e.code !== 'ENOENT') throw e;
         warn('WARN: tsConfig specified but not found:', upath.resolve(tsConfig));
       }
     }
-    if (!tsProject) {
-      tsProject = typescript.createProject(tsOpts);
-      if (builder.buildOptions.printConfig)
-        msg(`[TypeScriptPlugin]tsconfig evaluated(buildName:${builder.conf.buildName}):\n`,
-          {"compilerOptions":tsOpts});
+    if (!tsProject) tsProject = typescript.createProject(tsOpts);
+
+    if (builder.buildOptions.printConfig) {
+      msg(`[TypeScriptPlugin]tsconfig evaluated(buildName:${builder.conf.buildName}):\n`, tsProject.options);
     }
 
     // transpile .ts files